Extract coalescing cleanup ownership
